+++ Tests/Mockolate.Internal.Tests/Internals/TypeFormatterTests.cs
@@ -2,7 +2,7 @@
 using System.Reflection;
 using Mockolate.Internals;
 
-namespace Mockolate.Internal.Tests;
+namespace Mockolate.Internal.Tests.Internals;
 
 public sealed class TypeFormatterTests
 {
