+++ Tests/Mockolate.Internal.Tests/Setup/MockSetupsTests.IndexersTests.cs
@@ -1,11 +1,11 @@
 using Mockolate.Internal.Tests.TestHelpers;
 using Mockolate.Setup;
 
-namespace Mockolate.Internal.Tests;
+namespace Mockolate.Internal.Tests.Setup;
 
 public partial class MockSetupsTests
 {
-	public class IndexerSetupsTests
+	public class IndexersTests
 	{
 		[Fact]
 		public async Task AddAndGetLatestOrDefault_ShouldReturnLatestMatching()
@@ -17,8 +17,8 @@ public partial class MockSetupsTests
 			setups.Add(setup1);
 			setups.Add(setup2);
 
-			FakeIndexerSetup? result = setups.GetMatching<FakeIndexerSetup>(
-				s => ((IInteractiveIndexerSetup)s).Matches(access));
+			FakeIndexerSetup? result =
+				setups.GetMatching<FakeIndexerSetup>(s => ((IInteractiveIndexerSetup)s).Matches(access));
 
 			await That(result).IsEqualTo(setup1);
 		}
@@ -48,15 +48,14 @@ public partial class MockSetupsTests
 				FakeIndexerSetup setup = new(shouldMatch);
 				setups.Add(setup);
 				FakeIndexerAccess access = new();
-				_ = setups.GetMatching<FakeIndexerSetup>(
-					s => ((IInteractiveIndexerSetup)s).Matches(access));
+				_ = setups.GetMatching<FakeIndexerSetup>(s => ((IInteractiveIndexerSetup)s).Matches(access));
 			});
 			FakeIndexerSetup finalMatch = new(true);
 			setups.Add(finalMatch);
 			FakeIndexerAccess finalAccess = new();
 
-			FakeIndexerSetup? result = setups.GetMatching<FakeIndexerSetup>(
-				s => ((IInteractiveIndexerSetup)s).Matches(finalAccess));
+			FakeIndexerSetup? result =
+				setups.GetMatching<FakeIndexerSetup>(s => ((IInteractiveIndexerSetup)s).Matches(finalAccess));
 
 			await That(result).IsEqualTo(finalMatch);
 		}
